// Angular/Angular Intro With Login And Sidenav/src/app/pages/data-manipulation/data-bindings/enums.ts
export enum Operators {
    ADDITION = 'addition',
    SUBTRACTION = 'subtraction',
    MULTIPLICATION = 'multiplication',
    DIVISION = 'division',
}

export enum Strings {
    FIRST_NUMBER = 'firstNumber',
    SECOND_NUMBER = 'secondNumber',
    OPERATOR = 'operator',
}

export enum LifecycleHooks {
    ONINIT = 'OnInit',
    AFTERVIEWINIT = 'AfterViewInit',
    ONCHANGES = 'OnChanges',
    ONDESTROY = 'OnDestroy',
}